--- a/libgadget/forcetree.h
+++ b/libgadget/forcetree.h
@@ -18,6 +18,9 @@
 
 struct NodeChild
 {
+    /* Stores the types of the child particles. Uses a bit field, 3 bits per particle.
+     */
+    uint64_t Types;
     /*!< pointers to daughter nodes or daughter particles. */
     int suns[NMAXCHILD];
     /* Number of daughter particles if node contains particles.