--- a/libgadget/forcetree.h
+++ b/libgadget/forcetree.h
@@ -9,7 +9,7 @@
  */
 
 /* Total allowed number of particle children for a node*/
-#define NMAXCHILD 11
+#define NMAXCHILD 8
 
 /* Defines for the type of node, classified by type of children.*/
 #define PARTICLE_NODE_TYPE 0
@@ -20,7 +20,7 @@ struct NodeChild
 {
     /* Stores the types of the child particles. Uses a bit field, 3 bits per particle.
      */
-    uint64_t Types;
+    unsigned int Types;
     /*!< pointers to daughter nodes or daughter particles. */
     int suns[NMAXCHILD];
     /* Number of daughter particles if node contains particles.