Store the particle Type in the node structure so we can filter particles without loading them into memory. This seems to have a marginal effect on simulation speed.