@sbird 2020-09-10
Performance data (first 11 steps lyman alpha):
master:
Step 11, Time: 0.0106138, MPIs: 1 Threads: 8 Elapsed: 88.5753
 SPH                              18.35 20.7%        0.22 38.2%        0.22       0.22
  Density                          11.68 13.2%        0.20 35.0%        0.20       0.20
   Comm@density.c:0382               0.00  0.0%        0.00  0.0%        0.00       0.00
   Compute@density.c:0380            9.49 10.7%        0.02  2.8%        0.02       0.02
   Init@density.c:0278               1.92  2.2%        0.18 31.3%        0.18       0.18
   Misc@density.c:0383               0.27  0.3%        0.01  0.9%        0.01       0.01
   Wait@density.c:0381               0.00  0.0%        0.00  0.0%        0.00       0.00
  Hydro                             6.67  7.5%        0.02  3.2%        0.02       0.02
   Comm@hydra.c:0202                 0.00  0.0%        0.00  0.0%        0.00       0.00
   Compute@hydra.c:0200              6.61  7.5%        0.02  3.1%        0.02       0.02
   Misc@hydra.c:0203                 0.06  0.1%        0.00  0.1%        0.00       0.00
   Wait@hydra.c:0201                 0.00  0.0%        0.00  0.0%        0.00       0.00

this branch without the last commit (NMAXCHILD=11):
Step 11, Time: 0.0106138, MPIs: 1 Threads: 8 Elapsed: 88.0389
 SPH                              16.95 19.3%        0.21 38.1%        0.21       0.21
  Density                          10.93 12.4%        0.20 35.2%        0.20       0.20
   Comm@density.c:0382               0.00  0.0%        0.00  0.0%        0.00       0.00
   Compute@density.c:0380            8.69  9.9%        0.01  2.5%        0.01       0.01
   Init@density.c:0278               1.98  2.2%        0.18 31.7%        0.18       0.18
   Misc@density.c:0383               0.27  0.3%        0.01  1.0%        0.01       0.01
   Wait@density.c:0381               0.00  0.0%        0.00  0.0%        0.00       0.00
  Hydro                             6.02  6.8%        0.02  2.9%        0.02       0.02
   Comm@hydra.c:0202                 0.00  0.0%        0.00  0.0%        0.00       0.00
   Compute@hydra.c:0200              5.96  6.8%        0.02  2.8%        0.02       0.02
   Misc@hydra.c:0203                 0.06  0.1%        0.00  0.0%        0.00       0.00
   Wait@hydra.c:0201                 0.00  0.0%        0.00  0.0%        0.00       0.00

Full microopts:
Step 11, Time: 0.0106138, MPIs: 1 Threads: 8 Elapsed: 87.4987
 SPH                              17.19 19.6%        0.22 30.8%        0.22       0.22
  Density                          10.99 12.6%        0.21 28.6%        0.21       0.21
   Comm@density.c:0382               0.00  0.0%        0.00  0.0%        0.00       0.00
   Compute@density.c:0380            8.69  9.9%        0.01  1.8%        0.01       0.01
   Init@density.c:0278               2.03  2.3%        0.19 26.1%        0.19       0.19
   Misc@density.c:0383               0.27  0.3%        0.01  0.7%        0.01       0.01
   Wait@density.c:0381               0.00  0.0%        0.00  0.0%        0.00       0.00
  Hydro                             6.20  7.1%        0.02  2.2%        0.02       0.02
   Comm@hydra.c:0202                 0.00  0.0%        0.00  0.0%        0.00       0.00
   Compute@hydra.c:0200              6.14  7.0%        0.02  2.1%        0.02       0.02
   Misc@hydra.c:0203                 0.06  0.1%        0.00  0.0%        0.00       0.00
   Wait@hydra.c:0201                 0.00  0.0%        0.00  0.0%        0.00       0.00
(but it makes it up in the treewalk time)