@sbird 2020-09-10
So it is positive but fairly marginal.