@rainwoodman 2020-09-10
The speed up could be larger later in the simulation? I guess the number of cache misses should go down and the hotspots from line by line profiling shall shift after this PR?