@sbird 2020-09-10
I think it is probably true that it will be faster later in the simulation: it might also be a larger difference for more modern processors (this was my 4-core Haswell machine, which I think doesn't do AVX2). I haven't checked the line by line cache misses and hotspots yet. I think this is worth merging but I'm not sure further efforts along these lines are worth it.